requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
greenlet==3.0.3
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-dotenv==1.0.1
email-validator==2.2.0
pytest==8.3.3
httpx==0.27.2
python-multipart==0.0.20
openpyxl==3.1.5